src/test/java/com/assignment/spring/exception/GlobalExceptionHandlerTest.java
package com.assignment.spring.exception;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.client.HttpClientErrorException;
import org.springframework.web.client.RestClientException;
import javax.validation.ConstraintViolationException;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.mock;

@ExtendWith(MockitoExtension.class)
class GlobalExceptionHandlerTest {

    @InjectMocks
    private GlobalExceptionHandler globalExceptionHandler;

    @Test
    void handleRuntimeException() {
        RuntimeException ex = new RuntimeException("Internal Server Error");

        ResponseEntity<String> response = globalExceptionHandler.handleRuntimeException(ex);

        assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, response.getStatusCode());
        assertEquals("An error occurred: Internal Server Error", response.getBody());
    }

    @Test
    void handleRestClientException() {
        RestClientException ex = new RestClientException("Rest Client Error");

        ResponseEntity<ApiError> response = globalExceptionHandler.handleRuntimeException(ex);

        assertEquals(HttpStatus.INTERNAL_SERVER_ERROR, response.getStatusCode());
        assertEquals("An error occurred: Rest Client Error", response.getBody().getMessage());
    }

    @Test
    void handleClientNotFoundException() {
        HttpClientErrorException ex = mock(HttpClientErrorException.class);

        ResponseEntity<ApiError> response = globalExceptionHandler.handleClientNotFoundException(ex);

        assertEquals(HttpStatus.NOT_FOUND, response.getStatusCode());
    }

    @Test
    void handleConstraintViolationException() {
        ConstraintViolationException ex = new ConstraintViolationException("Invalid data", null);

        ResponseEntity<ApiError> response = globalExceptionHandler.handleConstraintViolationException(ex);

        assertEquals(HttpStatus.BAD_REQUEST, response.getStatusCode());
        assertEquals("Invalid data", response.getBody().getMessage());
    }
}
